refactor(graduate): lint 오류 수정
