+++ apps/graduate/src/features/notices/components/index.ts
@@ -1,0 +1,2 @@
+export { default as NoticeForm } from './NoticeForm';
+export { default as NoticesSection } from './NoticesSection';
