+++ apps/graduate/src/features/schedule/components/index.ts
@@ -1,0 +1,3 @@
+export { default as ScheduleSection } from './ScheduleSection';
+export { default as ScheduleEditModal } from './ScheduleEditModal';
+export { default as ScheduleDescription } from './ScheduleDescription';
