+++ apps/graduate/src/features/schedule/mock/schedule.ts
@@ -55,7 +55,7 @@ export const descriptionData = [
 		id: 5,
 		title: '최종 통과',
 		description:
-			`최종보고서는 지도교수 포함 3인 이상의 심사위원의 평가에 의해 통과 여부가 결정됩니다. 심사에서 통과된 경우 승인으로 표시됩니다.`.trim(),
+			'최종보고서는 지도교수 포함 3인 이상의 심사위원의 평가에 의해 통과 여부가 결정됩니다. 심사에서 통과된 경우 승인으로 표시됩니다.',
 	},
 	{
 		id: 6,
