+++ apps/graduate/src/routes/_adminLayout/notices/create.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import NoticeForm from '~/features/notices/components/NoticeForm';
+import { NoticeForm } from '~/features/notices/components';
 
 export const Route = createFileRoute('/_adminLayout/notices/create')({
 	component: RouteComponent,
