+++ apps/graduate/src/routes/_adminLayout/schedule.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import ScheduleSection from '~/features/schedule/components/ScheduleSection';
+import { ScheduleSection } from '~/features/schedule/components';
 
 export const Route = createFileRoute('/_adminLayout/schedule')({
 	component: RouteComponent,
